requirements.txt
python_dateutil==2.8.2
selenium==4.4.0
undetected_chromedriver==3.1.5.post4
